+++ includes/parsing.h
@@ -6,7 +6,7 @@
 /*   By: smagdela <smagdela@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/01 13:31:10 by smagdela          #+#    #+#             */
-/*   Updated: 2022/03/22 14:59:07 by smagdela         ###   ########.fr       */
+/*   Updated: 2022/03/22 17:32:53 by smagdela         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -77,7 +77,7 @@ bool	executor(t_token *token_list, char **envp, t_data *data);
 
 void	display_toklist(t_token *token_list);
 void	relink_toklist(t_token *elem, t_token *tmp,
-			char *new_data, t_token *token_list);
+			char *new_data, t_token **token_list);
 void	expand(t_token *elem, t_data *env_data);
 void	lst_pop(t_token *elem, t_token **token_list);
 void	reduce_all(t_token *elem, t_token *end, t_token *token_list);
@@ -85,7 +85,7 @@ void	reduce(t_token *elem, t_token *end, t_data *env_data,
 			t_token *token_list);
 void	ft_print_title(void);
 void	expand_remaining_envar(t_token *token_list, t_data *env_data);
-void	suppress_spaces(t_token *token_list);
+void	suppress_spaces(t_token **token_list);
 
 int		categ_1(t_input *input, t_token **token_list);
 int		categ_2(t_input *input, t_token **token_list);
