+++ srcs/analyzer.c
@@ -6,7 +6,7 @@
 /*   By: smagdela <smagdela@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/08 18:42:02 by smagdela          #+#    #+#             */
-/*   Updated: 2022/03/22 15:26:15 by smagdela         ###   ########.fr       */
+/*   Updated: 2022/03/22 17:39:37 by smagdela         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,27 +28,10 @@ static void	synerror(char *str, t_token **token_list)
 	}
 }
 
-t_token	*analyzer(t_token *token_list, t_data *env_data)
+static t_token	*last_check(t_token *token_list)
 {
 	t_token	*tmp;
 
-	if (checker_quotes(token_list, env_data) == false)
-	{
-		synerror(" : Near token quote.", &token_list);
-		return (token_list);
-	}
-	expand_remaining_envar(token_list, env_data);
-	suppress_spaces(token_list);
-	if (checker_words(token_list) == false)
-	{
-		synerror(" : Near token word.", &token_list);
-		return (token_list);
-	}
-	if (checker_redir(token_list) == false)
-	{
-		synerror(" : Near redirection token.", &token_list);
-		return (token_list);
-	}
 	tmp = token_list;
 	while (tmp != NULL)
 	{
@@ -61,5 +44,30 @@ t_token	*analyzer(t_token *token_list, t_data *env_data)
 		}
 		tmp = tmp->next;
 	}
+	return (token_list);
+}
+
+t_token	*analyzer(t_token *token_list, t_data *env_data)
+{
+	if (checker_quotes(token_list, env_data) == false)
+	{
+		synerror(" : Near token quote.", &token_list);
+		return (token_list);
+	}
+	expand_remaining_envar(token_list, env_data);
+	suppress_spaces(&token_list);
+	if (token_list == NULL)
+		return (NULL);
+	if (checker_words(token_list) == false)
+	{
+		synerror(" : Near token word.", &token_list);
+		return (token_list);
+	}
+	if (checker_redir(token_list) == false)
+	{
+		synerror(" : Near redirection token.", &token_list);
+		return (token_list);
+	}
+	token_list = last_check(token_list);
 	return (token_list);
 }
