--- srcs/categorizer_2.c
@@ -1,77 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   categorizer_2.c                                    :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: smagdela <smagdela@student.42.fr>          +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/02/08 21:42:47 by smagdela          #+#    #+#             */
-/*   Updated: 2022/02/11 17:23:09 by smagdela         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "parsing.h"
-
-static int	categorizer_aux(t_input *input, t_token **token_list)
-{
-	int	ret;
-
-	ret = categ_1(input, token_list);
-	if (ret != 1)
-		return (ret);
-	ret = categ_2(input, token_list);
-	if (ret != 1)
-		return (ret);
-	ret = categ_3(input, token_list);
-	if (ret != 1)
-		return (ret);
-	ret = categ_4(input, token_list);
-	if (ret != 1)
-		return (ret);
-	ret = categ_5(input, token_list);
-	if (ret != 1)
-		return (ret);
-	return (-1);
-}
-
-/*
-Ask what category is the data from the input, and create token accordingly.
-*/
-bool	categorizer(t_input *input, t_token **token_list)
-{
-	int	ret;
-
-	ret = categorizer_aux(input, token_list);
-	if (ret == 0)
-		return (true);
-	free_toklist(*token_list);
-	return (false);
-}
-
-/*
-Printing introduction banner...
-*/
-void	ft_print_title(void)
-{
-	ft_putstr_fd("\n\033[0;92m::::     :::: ::::::::::: ::::    :::", 1);
-	ft_putstr_fd(" ::::::::::: ::::::::  :::    ::: :::::::::: ::: ", 1);
-	ft_putstr_fd("       :::             :::             \n", 1);
-	ft_putstr_fd("+:+:+: :+:+:+     :+:     :+:+:   :+:     :+:    ", 1);
-	ft_putstr_fd(":+:    :+: :+:    :+: :+:        :+:        :+:  ", 1);
-	ft_putstr_fd("            :+:            \n", 1);
-	ft_putstr_fd("+:+ +:+:+ +:+     +:+     :+:+:+  +:+     +:+    ", 1);
-	ft_putstr_fd("+:+        +:+    +:+ +:+        +:+        +:+  ", 1);
-	ft_putstr_fd("             +:+           \n", 1);
-	ft_putstr_fd("+#+  +:+  +#+     +#+     +#+ +:+ +#+     +#+    ", 1);
-	ft_putstr_fd("+#++:++#++ +#++:++#++ +#++:++#   +#+        +#+  ", 1);
-	ft_putstr_fd("              +#+          \n", 1);
-	ft_putstr_fd("+#+       +#+     +#+     +#+  +#+#+#     +#+    ", 1);
-	ft_putstr_fd("       +#+ +#+    +#+ +#+        +#+        +#+  ", 1);
-	ft_putstr_fd("	       +#+            \n", 1);
-	ft_putstr_fd("#+#       #+#     #+#     #+#   #+#+#     #+#    ", 1);
-	ft_putstr_fd("#+#    #+# #+#    #+# #+#        #+#        #+#  ", 1);
-	ft_putstr_fd("            #+#            \n", 1);
-	ft_putstr_fd("###       ### ########### ###    #### ###########", 1);
-	ft_putstr_fd(" ########  ###    ### ########## ########## #####", 1);
-	ft_putstr_fd("#####      ###    ##########\033[0;m\n\n", 1);
-}
