+++ srcs/scanner.c
@@ -6,7 +6,7 @@
 /*   By: smagdela <smagdela@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/02 18:31:12 by smagdela          #+#    #+#             */
-/*   Updated: 2022/03/22 15:29:26 by smagdela         ###   ########.fr       */
+/*   Updated: 2022/03/22 17:26:35 by smagdela         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,32 +31,40 @@ bool	is_in_charset(char c, char *charset)
 	return (false);
 }
 
-/*
-Completely frees the token list, as well as data for each concerned token.
-Returns -1 (for norm and ease of code reasons).
-*/
-int	free_toklist(t_token *list)
+static int	categorizer_aux(t_input *input, t_token **token_list)
 {
-	t_token	*tmp;
-	t_token	*to_free;
+	int	ret;
 
-	if (list != NULL)
-	{
-		tmp = list;
-		while (tmp)
-		{
-			if (tmp->type == WORD || tmp->type == VAR)
-				free(tmp->data);
-			if (tmp->cmd != NULL)
-				free_tab(tmp->cmd);
-			to_free = tmp;
-			tmp = tmp->next;
-			free(to_free);
-			printf("Trying to free the %luth token, of data [%s]\n", tmp->index, tmp->data);
-			to_free = NULL;
-		}
-	}
+	ret = categ_1(input, token_list);
+	if (ret != 1)
+		return (ret);
+	ret = categ_2(input, token_list);
+	if (ret != 1)
+		return (ret);
+	ret = categ_3(input, token_list);
+	if (ret != 1)
+		return (ret);
+	ret = categ_4(input, token_list);
+	if (ret != 1)
+		return (ret);
+	ret = categ_5(input, token_list);
+	if (ret != 1)
+		return (ret);
 	return (-1);
+}
+
+/*
+Ask what category is the data from the input, and create token accordingly.
+*/
+bool	categorizer(t_input *input, t_token **token_list)
+{
+	int	ret;
+
+	ret = categorizer_aux(input, token_list);
+	if (ret == 0)
+		return (true);
+	free_toklist(*token_list);
+	return (false);
 }
 
 /*
@@ -69,7 +77,6 @@ t_token	*scanner(const char *str)
 
 	if (!str)
 	{
-		token_list = NULL;
 		ft_putstr_fd("Error: Null string passed to scanner\n", 2);
 		return (NULL);
 	}
@@ -81,7 +88,33 @@ t_token	*scanner(const char *str)
 		if (categorizer(&input, &token_list) == false)
 			break ;
 	}
-	if (token_list == NULL)
-		return (NULL);
 	return (token_list);
+}
+
+/*
+Printing introduction banner...
+*/
+void	ft_print_title(void)
+{
+	ft_putstr_fd("\n\033[0;92m::::     :::: ::::::::::: ::::    :::", 1);
+	ft_putstr_fd(" ::::::::::: ::::::::  :::    ::: :::::::::: ::: ", 1);
+	ft_putstr_fd("       :::             :::             \n", 1);
+	ft_putstr_fd("+:+:+: :+:+:+     :+:     :+:+:   :+:     :+:    ", 1);
+	ft_putstr_fd(":+:    :+: :+:    :+: :+:        :+:        :+:  ", 1);
+	ft_putstr_fd("            :+:            \n", 1);
+	ft_putstr_fd("+:+ +:+:+ +:+     +:+     :+:+:+  +:+     +:+    ", 1);
+	ft_putstr_fd("+:+        +:+    +:+ +:+        +:+        +:+  ", 1);
+	ft_putstr_fd("             +:+           \n", 1);
+	ft_putstr_fd("+#+  +:+  +#+     +#+     +#+ +:+ +#+     +#+    ", 1);
+	ft_putstr_fd("+#++:++#++ +#++:++#++ +#++:++#   +#+        +#+  ", 1);
+	ft_putstr_fd("              +#+          \n", 1);
+	ft_putstr_fd("+#+       +#+     +#+     +#+  +#+#+#     +#+    ", 1);
+	ft_putstr_fd("       +#+ +#+    +#+ +#+        +#+        +#+  ", 1);
+	ft_putstr_fd("	       +#+            \n", 1);
+	ft_putstr_fd("#+#       #+#     #+#     #+#   #+#+#     #+#    ", 1);
+	ft_putstr_fd("#+#    #+# #+#    #+# #+#        #+#        #+#  ", 1);
+	ft_putstr_fd("            #+#            \n", 1);
+	ft_putstr_fd("###       ### ########### ###    #### ###########", 1);
+	ft_putstr_fd(" ########  ###    ### ########## ########## #####", 1);
+	ft_putstr_fd("#####      ###    ##########\033[0;m\n\n", 1);
 }
